refactor: ability detail fragment 연결
